requirements.txt
streamlit
llama-cpp-python
python-docx
PyMuPDF
